+++ commandCenter.go
@@ -21,6 +21,7 @@ type CommandCenter struct {
 	EtcdClient *clientv3.Client
 	ID         string
 	Nick       string
+	Team       string
 	StealData  struct {
 		LastAttackTime time.Time
 		AttackInterval time.Duration
@@ -31,10 +32,17 @@ type CommandCenter struct {
 type State struct {
 	ID    string `json:"id"`
 	Nick  string `json:"nick"`
+	Team  string `json:"team"`
 	Funds struct {
 		Amount float32 `json:"amount"`
 		*sync.RWMutex
 	} `json:"funds"`
+	Vault struct {
+		Amount   float32 `json:"amount"`
+		Level    float32 `json:"level"`
+		Capacity float32 `json:"capacity"`
+		*sync.RWMutex
+	} `json:"vault"`
 	Firewall struct {
 		Level float32 `json:"level"`
 	} `json:"firewall"`
@@ -116,14 +124,15 @@ func (c *CommandCenter) UpdateStealList(stealerName string, stolenAmount float32
 }
 
 // Set Steal Cooldown
-// Cap the cooldown to 24 hours max
-func (c *CommandCenter) SetCooldown(levelDifference int) {
-	attackCooldown := 5 + levelDifference
-	if time.Minute*time.Duration(attackCooldown) > time.Hour*time.Duration(24) {
-		c.StealData.AttackInterval = time.Hour * time.Duration(24)
-	} else {
-		c.StealData.AttackInterval = time.Minute * time.Duration(attackCooldown)
-	}
+// Set to a fixed time of 10 Minutes after a steal
+func (c *CommandCenter) SetCooldown() {
+	// attackCooldown := 5 + levelDifference
+	// if time.Minute*time.Duration(attackCooldown) > time.Hour*time.Duration(24) {
+	// 	c.StealData.AttackInterval = time.Hour * time.Duration(24)
+	// } else {
+	// 	c.StealData.AttackInterval = time.Minute * time.Duration(attackCooldown)
+	// }
+	c.StealData.AttackInterval = time.Minute * 10
 	c.State.CoolDown.ExpirationTimeStamp = int64(time.Now().Add(c.StealData.AttackInterval).Unix())
 }
 
@@ -132,17 +141,24 @@ func (c *CommandCenter) Init(config clientv3.Config) CommandCenter {
 	// Configure commandCenter State
 	c.ID = getEnv("ID", "123456")
 	c.Nick = getEnv("NICK", "DEBUGPLAYER")
+	c.Team = getEnv("TEAM", "none")
 	c.State.ID = c.ID
 	c.State.Nick = c.Nick
+	c.State.Team = c.Team
 	c.State.CryptoMiner.Level = 1
 	c.State.Scanner.Level = 1
 	c.State.Firewall.Level = 1
 	c.State.Stealer.Level = 1
+	c.State.Vault.Level = 1
+	c.State.Vault.Capacity = 10
 	// Set initial protection period on boot
 	c.StealData.LastAttackTime = time.Now()
 	// Configure starting attack interval
 	c.StealData.AttackInterval = time.Minute * 5
+	// Set initial gui cooldown time
+	c.State.CoolDown.ExpirationTimeStamp = int64(time.Now().Add(c.StealData.AttackInterval).Unix())
 	c.State.Funds.RWMutex = &sync.RWMutex{}
+	c.State.Vault.RWMutex = &sync.RWMutex{}
 	c.State.LastSteals = []StealData{}
 
 	// Init etcd client
@@ -177,6 +193,40 @@ func (c *CommandCenter) Init(config clientv3.Config) CommandCenter {
 	return *c
 }
 
+// Store coins in vault
+func (c *CommandCenter) StoreVault() float32 {
+	c.State.Vault.Lock()
+	c.State.Funds.Lock()
+	defer c.State.Vault.Unlock()
+	defer c.State.Funds.Unlock()
+	// Transfer money to vault but do not exceed the max capacity.
+	transferAmount := c.State.Funds.Amount
+	// Check if adding the current liquid funds would exceed the capacity of the vault.
+	if c.State.Vault.Amount+transferAmount > c.State.Vault.Capacity {
+		amountToAdd := c.State.Vault.Capacity - c.State.Vault.Amount
+		c.State.Vault.Amount = c.State.Vault.Capacity
+		c.State.Funds.Amount -= amountToAdd
+		return amountToAdd
+	}
+	c.State.Vault.Amount += transferAmount
+	// Empty the Funds
+	c.State.Funds.Amount = 0
+	return transferAmount
+}
+
+// Store coins in vault
+func (c *CommandCenter) WithdrawVault() float32 {
+	c.State.Vault.Lock()
+	c.State.Funds.Lock()
+	defer c.State.Vault.Unlock()
+	defer c.State.Funds.Unlock()
+	// Withdraw all coins from vault and set vault amount to 0
+	transferAmount := c.State.Vault.Amount
+	c.State.Funds.Amount += transferAmount
+	c.State.Vault.Amount = 0
+	return transferAmount
+}
+
 // Get value from etcd3
 func (c *CommandCenter) GetValue(key string) []byte {
 	ctx, cancel := context.WithTimeout(context.Background(), 2*time.Second)
@@ -199,7 +249,7 @@ func (c *CommandCenter) PutValue(key string, value string) {
 	cancel()
 	//fmt.Printf("Put value %s on topic: %s\n", value, key)
 	if err != nil {
-		log.Panicln("Could not put key")
+		log.Println("Could not put key")
 	}
 }
 
@@ -209,7 +259,7 @@ func (c *CommandCenter) DeleteState(key string) error {
 	_, err := c.EtcdClient.Delete(ctx, key)
 	cancel()
 	if err != nil {
-		log.Panicln("Could not delete key")
+		log.Println("Could not delete key")
 		return err
 	}
 	return nil
@@ -235,13 +285,13 @@ func (c *CommandCenter) SaveStateContinuous() {
 	for range ticker.C {
 		value, marshalErr := json.Marshal(c.State)
 		if marshalErr != nil {
-			log.Panicf("Could not save state.\n")
+			log.Println("Could not save state.")
 		}
 		ctx, cancel := context.WithTimeout(context.Background(), 2*time.Second)
 		_, err := c.EtcdClient.Put(ctx, c.ID, string(value))
 		cancel()
 		if err != nil {
-			log.Panicln("Could not put key")
+			log.Println("Could not put key. Trying again in a minute")
 		}
 	}
 }
@@ -250,13 +300,13 @@ func (c *CommandCenter) SaveStateContinuous() {
 func (c *CommandCenter) SaveStateImmediate() {
 	value, marshalErr := json.Marshal(c.State)
 	if marshalErr != nil {
-		log.Panicf("Could not save state.\n")
+		log.Println("Could not save state.")
 	}
 	ctx, cancel := context.WithTimeout(context.Background(), 2*time.Second)
 	_, err := c.EtcdClient.Put(ctx, c.ID, string(value))
 	cancel()
 	if err != nil {
-		log.Panicln("Could not put key")
+		log.Println("Could not put key.")
 	}
 
 }
@@ -269,16 +319,47 @@ func (c *CommandCenter) Mine(wg *sync.WaitGroup) {
 		c.State.Funds.Lock()
 		c.State.Funds.Amount += baseMineRate * float32(minerLevel)
 		c.State.Funds.Unlock()
-		monitor.MinedCoins.WithLabelValues(c.ID, c.Nick).Add(float64(baseMineRate * float32(minerLevel)))
+		monitor.MinedCoins.WithLabelValues(c.ID, c.Nick, c.Team).Add(float64(baseMineRate * float32(minerLevel)))
 		time.Sleep(time.Second * 1)
 	}
+}
+
+// UpgradeVault on CommandCenter this will not have a max buy option since the vault upgrade costs as much as its capacity
+func (c *CommandCenter) UpgradeVault(nc *nats.Conn, upgradeToMax bool) (bool, *CommandCenter, UpgradeReply, error) {
+	reply := UpgradeReply{}
+	c.State.Vault.Lock()
+	defer c.State.Vault.Unlock()
+	state, err := json.Marshal(&c.State)
+	if err != nil {
+		log.Fatalln(err)
+		return false, c, reply, err
+	}
+	msg, err := nc.Request(fmt.Sprintf("commandcenter.%s.upgradeVault", c.ID), []byte(state), time.Second)
+	if err != nil {
+		log.Fatalln(err)
+		return false, c, reply, err
+	}
+	log.Printf("UpdateVault reply: %s", msg.Data)
+	jsonErr := json.Unmarshal(msg.Data, &reply)
+	if jsonErr != nil {
+		log.Fatalln(jsonErr)
+		return false, c, reply, err
+	}
+	if reply.Allow {
+		c.State.Vault.Amount = c.State.Vault.Amount - reply.Cost
+		monitor.SpentCoins.WithLabelValues(c.ID, c.Nick, c.Team).Add(float64(reply.Cost))
+		c.State.Vault.Level += reply.Levels
+		c.State.Vault.Capacity += 10 * reply.Levels
+		return true, c, reply, err
+	}
+	return false, c, reply, err
 }
 
 // UpgradeStealer on CommandCenter
 func (c *CommandCenter) UpgradeStealer(nc *nats.Conn, upgradeToMax bool) (bool, *CommandCenter, UpgradeReply, error) {
 	reply := UpgradeReply{}
-	c.State.Funds.Lock()
-	defer c.State.Funds.Unlock()
+	c.State.Vault.Lock()
+	defer c.State.Vault.Unlock()
 	state, err := json.Marshal(&c.State)
 	if err != nil {
 		log.Fatalln(err)
@@ -301,8 +382,8 @@ func (c *CommandCenter) UpgradeStealer(nc *nats.Conn, upgradeToMax bool) (bool, 
 		return false, c, reply, err
 	}
 	if reply.Allow {
-		c.State.Funds.Amount = c.State.Funds.Amount - reply.Cost
-		monitor.SpentCoins.WithLabelValues(c.ID, c.Nick).Add(float64(reply.Cost))
+		c.State.Vault.Amount = c.State.Vault.Amount - reply.Cost
+		monitor.SpentCoins.WithLabelValues(c.ID, c.Nick, c.Team).Add(float64(reply.Cost))
 		c.State.Stealer.Level += reply.Levels
 		return true, c, reply, err
 	}
@@ -312,8 +393,8 @@ func (c *CommandCenter) UpgradeStealer(nc *nats.Conn, upgradeToMax bool) (bool, 
 // UpgradeFirewall on CommandCenter
 func (c *CommandCenter) UpgradeFirewall(nc *nats.Conn, upgradeToMax bool) (bool, *CommandCenter, UpgradeReply, error) {
 	reply := UpgradeReply{}
-	c.State.Funds.Lock()
-	defer c.State.Funds.Unlock()
+	c.State.Vault.Lock()
+	defer c.State.Vault.Unlock()
 	state, err := json.Marshal(&c.State)
 	if err != nil {
 		log.Fatalln(err)
@@ -336,8 +417,8 @@ func (c *CommandCenter) UpgradeFirewall(nc *nats.Conn, upgradeToMax bool) (bool,
 		return false, c, reply, err
 	}
 	if reply.Allow {
-		c.State.Funds.Amount = c.State.Funds.Amount - reply.Cost
-		monitor.SpentCoins.WithLabelValues(c.ID, c.Nick).Add(float64(reply.Cost))
+		c.State.Vault.Amount = c.State.Vault.Amount - reply.Cost
+		monitor.SpentCoins.WithLabelValues(c.ID, c.Nick, c.Team).Add(float64(reply.Cost))
 		c.State.Firewall.Level += reply.Levels
 		return true, c, reply, err
 	}
@@ -347,8 +428,8 @@ func (c *CommandCenter) UpgradeFirewall(nc *nats.Conn, upgradeToMax bool) (bool,
 // UpgradeScanner on CommandCenter
 func (c *CommandCenter) UpgradeScanner(nc *nats.Conn, upgradeToMax bool) (bool, *CommandCenter, UpgradeReply, error) {
 	reply := UpgradeReply{}
-	c.State.Funds.Lock()
-	defer c.State.Funds.Unlock()
+	c.State.Vault.Lock()
+	defer c.State.Vault.Unlock()
 	state, err := json.Marshal(&c.State)
 	if err != nil {
 		log.Fatalln(err)
@@ -371,8 +452,8 @@ func (c *CommandCenter) UpgradeScanner(nc *nats.Conn, upgradeToMax bool) (bool, 
 		return false, c, reply, err
 	}
 	if reply.Allow {
-		c.State.Funds.Amount = c.State.Funds.Amount - reply.Cost
-		monitor.SpentCoins.WithLabelValues(c.ID, c.Nick).Add(float64(reply.Cost))
+		c.State.Vault.Amount = c.State.Vault.Amount - reply.Cost
+		monitor.SpentCoins.WithLabelValues(c.ID, c.Nick, c.Team).Add(float64(reply.Cost))
 		c.State.Scanner.Level += reply.Levels
 		return true, c, reply, err
 	}
@@ -382,8 +463,8 @@ func (c *CommandCenter) UpgradeScanner(nc *nats.Conn, upgradeToMax bool) (bool, 
 // UpgradeCryptoMiner on CommandCenter
 func (c *CommandCenter) UpgradeCryptoMiner(nc *nats.Conn, upgradeToMax bool) (bool, *CommandCenter, UpgradeReply, error) {
 	reply := UpgradeReply{}
-	c.State.Funds.Lock()
-	defer c.State.Funds.Unlock()
+	c.State.Vault.Lock()
+	defer c.State.Vault.Unlock()
 	state, err := json.Marshal(&c.State)
 	if err != nil {
 		log.Fatalln(err)
@@ -406,8 +487,8 @@ func (c *CommandCenter) UpgradeCryptoMiner(nc *nats.Conn, upgradeToMax bool) (bo
 		return false, c, reply, err
 	}
 	if reply.Allow {
-		c.State.Funds.Amount = c.State.Funds.Amount - reply.Cost
-		monitor.SpentCoins.WithLabelValues(c.ID, c.Nick).Add(float64(reply.Cost))
+		c.State.Vault.Amount = c.State.Vault.Amount - reply.Cost
+		monitor.SpentCoins.WithLabelValues(c.ID, c.Nick, c.Team).Add(float64(reply.Cost))
 		c.State.CryptoMiner.Level += reply.Levels
 		return true, c, reply, nil
 	}
@@ -438,7 +519,7 @@ func (c *CommandCenter) ReplyScan(nc *nats.Conn) error {
 				// gpt71 countermeasure
 				if foreignCommandCenter.ID == "180887" {
 					//tmpState.CoolDown.Time += time.Duration(time.Duration(rand.Intn(10)*100 + 10).Seconds())
-					tmpState.CoolDown.ExpirationTimeStamp -= rand.Int63n(1000)
+					tmpState.CoolDown.ExpirationTimeStamp -= rand.Int63n(100)
 				}
 
 				jsonReply, err := json.Marshal(tmpState) //json.Marshal(c.State)
@@ -469,14 +550,14 @@ func (c *CommandCenter) ReplyScan(nc *nats.Conn) error {
 func (c *CommandCenter) RequestScan(nc *nats.Conn) ([]State, string, error) {
 	// This costs money, so we remove coins based on the level of the scanner. First check if enough funds are available for scan.
 	cost := c.State.Scanner.Level * 0.1
-	c.State.Funds.Lock()
-	defer c.State.Funds.Unlock()
-	if cost > c.State.Funds.Amount {
+	c.State.Vault.Lock()
+	defer c.State.Vault.Unlock()
+	if cost > c.State.Vault.Amount {
 		err := fmt.Errorf("not enough funds. cost: %f", cost)
 		return nil, fmt.Sprintf("Not enough funds. Cost: %f", cost), err
 	} else {
-		c.State.Funds.Amount = c.State.Funds.Amount - cost
-		monitor.SpentCoins.WithLabelValues(c.ID, c.Nick).Add(float64(cost))
+		c.State.Vault.Amount = c.State.Vault.Amount - cost
+		monitor.SpentCoins.WithLabelValues(c.ID, c.Nick, c.Team).Add(float64(cost))
 	}
 
 	var states = []State{}
@@ -605,7 +686,8 @@ func (c *CommandCenter) ReplySteal(nc *nats.Conn) error {
 				// attackCooldown := 5 + lvldiff
 				// c.StealData.AttackInterval = time.Minute * time.Duration(attackCooldown)
 				// Set the cooldown for new steals. Capped at 24 hours max.
-				c.SetCooldown(lvldiff)
+				// UPDATE: 2024-04-30: Set to a fixed 10 minutes to allow for a more dynamic gameplay
+				c.SetCooldown()
 				// Level of stealer is higher
 
 				// Current Funds = 5.11
@@ -615,16 +697,15 @@ func (c *CommandCenter) ReplySteal(nc *nats.Conn) error {
 				// Do not remove targetfunds if that would mean sinking below 0
 
 				// Create a fixed size array based on the level difference
-				stealAttempts := make([]float32, lvldiff)
+				stealAttempts := make([]int, lvldiff)
 				// Create a coin cache
 				var coincache float32
 				// Dyncamically determine amount to steal per attempt by level.
-				stealAmount := float32(0.01) * stealerLevel
+				stealAmount := float32(0.011) * stealerLevel
 				log.Printf("Coins are being stolen! %f per attempt!", stealAmount)
 				// Try to steal money on each iteration. Logic is described on top.
 				c.State.Funds.Lock()
-				for i := range stealAttempts {
-					stealAttempts[i] = float32(i)
+				for range stealAttempts {
 					//  futureFunds := c.State.Funds.Amount - stealAmount; futureFunds > c.State.Funds.Amount
 					if c.State.Funds.Amount-stealAmount <= 0 {
 						// Player loses no money because firewall level changed mid loop.
@@ -646,7 +727,7 @@ func (c *CommandCenter) ReplySteal(nc *nats.Conn) error {
 				c.State.Funds.Unlock()
 				// Add stolen amount to personal steal archival list
 				c.UpdateStealList(foreignCommandCenter.Nick, coincache)
-				monitor.LostCoins.WithLabelValues(c.ID, c.Nick, foreignCommandCenter.ID, foreignCommandCenter.Nick).Add(float64(coincache))
+				monitor.LostCoins.WithLabelValues(c.ID, c.Nick, foreignCommandCenter.ID, foreignCommandCenter.Nick, foreignCommandCenter.Team).Add(float64(coincache))
 				// After successfully stealing from the target, send a reply to the attacker
 				stealReply := StealReply{
 					Attacker: struct {
@@ -684,16 +765,18 @@ func (c *CommandCenter) ReplySteal(nc *nats.Conn) error {
 // Here we want to start a steal event. Compared to the scan, this only targets specific command centers by ID.
 func (c *CommandCenter) StealFromTarget(targetId string, nc *nats.Conn) (StealReply, string, error) {
 	var stealReply StealReply
+	c.State.Vault.Lock()
 	c.State.Funds.Lock()
+	defer c.State.Vault.Unlock()
 	defer c.State.Funds.Unlock()
 	// This costs money, so we remove coins based on the level of the stealer. First check if enough funds are available for scan.
 	cost := c.State.Stealer.Level * 0.1
-	if cost > c.State.Funds.Amount {
+	if cost > c.State.Vault.Amount {
 		err := fmt.Errorf("not enough funds. cost: %f", cost)
 		return stealReply, fmt.Sprintf("Not enough funds. Cost: %f", cost), err
 	} else {
-		c.State.Funds.Amount = c.State.Funds.Amount - cost
-		monitor.SpentCoins.WithLabelValues(c.ID, c.Nick).Add(float64(cost))
+		c.State.Vault.Amount = c.State.Vault.Amount - cost
+		monitor.SpentCoins.WithLabelValues(c.ID, c.Nick, c.Team).Add(float64(cost))
 	}
 	state, err := json.Marshal(&c.State)
 	if err != nil {
@@ -725,7 +808,7 @@ func (c *CommandCenter) StealFromTarget(targetId string, nc *nats.Conn) (StealRe
 	sub.Unsubscribe()
 	// Add coins to account. If no coins are gained, then nothing will be added.
 	c.State.Funds.Amount += stealReply.GainedCoins
-	monitor.StolenCoins.WithLabelValues(c.ID, c.Nick).Add(float64(stealReply.GainedCoins))
+	monitor.StolenCoins.WithLabelValues(c.ID, c.Nick, c.Team).Add(float64(stealReply.GainedCoins))
 	// Return reply with error on cooldown
 	if stealReply.CoolDown {
 		err := fmt.Errorf("target is on cooldown")
@@ -734,21 +817,6 @@ func (c *CommandCenter) StealFromTarget(targetId string, nc *nats.Conn) (StealRe
 
 	return stealReply, "successful steal operation", nil
 }
-
-// Periodically update the ticker
-// func (c *CommandCenter) UpdateCoolDown() {
-// 	ticker := time.NewTicker(1 * time.Second)
-// 	for {
-// 		<-ticker.C
-// 		if c.StealData.AttackInterval/time.Second-(time.Since(c.StealData.LastAttackTime)/time.Second) > 0 {
-// 			c.State.CoolDown.Time = c.StealData.AttackInterval/time.Second - (time.Since(c.StealData.LastAttackTime) / time.Second)
-// 			monitor.CoolDown.WithLabelValues(c.ID, c.Nick).Set(float64(c.State.CoolDown.Time))
-// 		} else {
-// 			c.State.CoolDown.Time = 0
-// 			monitor.CoolDown.WithLabelValues(c.ID, c.Nick).Set(float64(c.State.CoolDown.Time))
-// 		}
-// 	}
-// }
 
 // Get state of CommandCenter
 func (c *CommandCenter) GetState() *CommandCenter {