+++ CHANGELOG.md
@@ -1,3 +1,30 @@
+# 1.1.1
+
+## Changes
+
+- [MECHANICS] Cooldowns are now always 10 Minutes. This should make upgrading the firewall more important.
+- [MECHANICS] Vault is now the primary source of funds for upgrades, etc. Transfers are essential to be able to buy upgrades. Coins in the vault are safe from theft. Coins not in the vault on the other hand, not.
+- [MECHANICS] Steal amount increased 10% | (0.1 -> 0.11) * stealerLevel
+
+# 1.1
+
+## Additions
+
+- [MECHANICS] Vault: The Vault allows you to store (for a 10% fee) all of your current funds. You can then later withdraw your funds to pay for upgrades. The Vault cannot be stolen from. The Funds in the Vault cannot be used to buy upgrades. They are frozen until withdrawn.
+
+# 1.0.3
+
+## Additions
+
+- [MECHANICS] "Buy Max" option is now available per component
+
+
+# 1.0.2
+
+## Additions
+
+- [MECHANICS] State now stores a list of the 50 last people who stole from you.
+
 # 1.0.1
 
 ## Changes